+++ server/src/components/MainFacade.js
@@ -1,0 +1,17 @@
+const express = require('express')
+const router = require('../routes')
+const cors = require('cors')
+const PORT = 3333
+const HOST = '0.0.0.0'
+
+module.exports = {
+    startServer() {
+        const app = express()
+        app.use(cors())
+        app.use(express.json())
+        app.use(router)
+        app.listen(PORT, HOST, () => {
+            console.log(`Server started at ${PORT}`)
+        })
+    }
+}
