+++ server/src/main.js
@@ -1,15 +1,3 @@
-const express = require('express')
-const router = require('./routes')
-const cors = require('cors')
-const PORT = 3333
-const HOST = '0.0.0.0'
+const mainFacade = require('./components/MainFacade');
 
-const app = express()
-
-app.use(cors())
-app.use(express.json())
-app.use(router)
-
-app.listen(PORT, HOST, () => {
-    console.log(`Server started at ${PORT}`)
-})
+mainFacade.startServer();
